@@ -2,7 +2,8 @@
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-//+build generate
+//go:build generate
+// +build generate
 
 package main
 
